Separating batch sampler and loader
