+++ dwi_ml/training/batch_samplers.py
@@ -1,32 +1,8 @@
 # -*- coding: utf-8 -*-
-from collections import defaultdict
-import logging
-from typing import Dict, List, Union, Tuple, Iterator
-
-from dipy.io.stateful_tractogram import StatefulTractogram
-import numpy as np
-from scilpy.tracking.tools import resample_streamlines_step_size
-from scilpy.utils.streamlines import compress_sft
-import torch
-import torch.multiprocessing
-from torch.utils.data import Sampler
-
-from dwi_ml.experiment_utils.prints import TqdmLoggingHandler
-from dwi_ml.data.dataset.multi_subject_containers import MultisubjectSubset
-from dwi_ml.data.processing.streamlines.data_augmentation import (
-    add_noise_to_streamlines, reverse_streamlines, split_streamlines)
-
-# For the batch sampler with inputs
-from dwi_ml.data.processing.volume.interpolation import \
-    interpolate_volume_in_neighborhood
-
 """
                                 Batch sampler
 
-These classes defines how to sample the streamlines available in the
-MultiSubjectData.
-
-AbstractBatchSampler:
+Defines how to sample the streamlines available in the MultiSubjectData.
 
 - Defines the __iter__ method:
     - Finds a list of streamlines ids and associated subj that you can later
@@ -37,28 +13,6 @@ AbstractBatchSampler:
     reduce the number of time we need to load new data by using the same
     subjects for a given number of "cycles".
 
-- Define the load_batch method:
-    - Loads the streamlines associated to sampled ids. Can resample them.
-
-    - Performs data augmentation (on-the-fly to avoid having to multiply data
-     on disk) (ex: splitting, reversing, adding noise).
-
-    NOTE: Actual loaded batch size might be different than `batch_size`
-    depending on chosen data augmentation. This sampler takes streamline
-    cutting and resampling into consideration, and as such, will return more
-    (or less) points than the provided `batch_size`.
-
-----------
-                        Implemented child classes
-
-BatchStreamlinesSamplerOneInput:
-
-- Redefines the load_batch method:
-    - Now also loads the input data under each point of the streamline (and
-    possibly its neighborhood), for one input volume.
-
-You are encouraged to contribute to dwi_ml by adding any child class here.
-
 USAGE:
 Can be used in a torch DataLoader. For instance:
         # Initialize dataset
@@ -68,19 +22,27 @@ Can be used in a torch DataLoader. For instance:
         batch_sampler = BatchSampler(...)
         # Use this in the dataloader
         dataloader = DataLoader(dataset, batch_sampler=batch_sampler,
-                                collate_fn=batch_sampler.load_batch)
+                                collate_fn=batch_loader.load_batch)
 """
 
+import logging
+from typing import List, Tuple, Iterator
 
-class AbstractBatchSampler(Sampler):
+import numpy as np
+import torch
+import torch.multiprocessing
+from torch.utils.data import Sampler
+
+from dwi_ml.experiment_utils.prints import TqdmLoggingHandler
+from dwi_ml.data.dataset.multi_subject_containers import MultisubjectSubset
+
+
+class DWIMLBatchSampler(Sampler):
     def __init__(self, dataset: MultisubjectSubset,
                  streamline_group_name: str, max_batch_size: int,
                  batch_size_unit: str, max_chunk_size: int, rng: int,
                  nb_subjects_per_batch: int, cycles: int,
-                 step_size: float, compress: bool,
-                 split_ratio: float, noise_gaussian_size: float,
-                 noise_gaussian_variability: float,
-                 reverse_ratio: float):
+                 step_size: float, compress: bool):
         """
         Parameters
         ----------
@@ -89,7 +51,8 @@ class AbstractBatchSampler(Sampler):
         streamline_group_name: str
             The name of the group to use to load the sequences among all
             streamline_groups in the data_source.
-        max_chunk_size: Number of streamlines to sample together while creating the
+        max_chunk_size: int
+            Number of streamlines to sample together while creating the
             batches. If the size of the streamlines is known in terms of
             number of points (resampling has been done, and no compressing is
             done in the batch sampler), we iteratively add chunk_size
@@ -126,37 +89,6 @@ class AbstractBatchSampler(Sampler):
             If true, compress streamlines. Cannot be used together with
             step_size. Once again, the choice can be different in the batch
             sampler than chosen when creating the hdf5.
-        split_ratio : float
-            DATA AUGMENTATION: Percentage of streamlines to randomly split
-            into 2, in each batch (keeping both segments as two independent
-            streamlines).
-            The reason for cutting is to help the ML algorithm to track from
-            the middle of WM by having already seen half-streamlines. If you
-            are using interface seeding, this is not necessary.
-        noise_gaussian_size : float
-            DATA AUGMENTATION: Add random Gaussian noise to streamline
-            coordinates with given variance. This corresponds to the std of the
-            Gaussian. If step_size is not given, make sure it is smaller than
-            your step size to avoid flipping direction. Ex, you could choose
-            0.1 * step-size. Noise is truncated to +/- 2*noise_sigma and to
-            +/- 0.5 * step-size (if given).
-        noise_gaussian_variability: float
-            DATA AUGMENTATION: If this is given, a variation is applied to the
-            streamline_noise_gaussian_size to have more noisy streamlines and
-            less noisy streamlines. This means that the real gaussian_size will
-            be a random number between [size - variability, size + variability]
-        reverse_ratio: float
-            DATA AUGMENTATION: If set, reversed a part of the streamlines in
-            the batch. You could want to reverse ALL your data and then use
-            both the initial data and reversed data. But this would take twice
-            the memory. If your ratio is, say, 0.5, streamlines have a 50%
-            chance to be reversed. If you train for enough epochs, high chance
-            that you will have used both directions of your streamline at least
-            once. Default: 0.5.
-            A way to absolutely ensure using both directions the same number of
-            time, we could use a flag and at each epoch, reverse those with
-            unreversed flag. But that adds a bool for each streamline in your
-            dataset and probably not so useful.
         """
         super().__init__(dataset)  # This does nothing but python likes it.
 
@@ -200,16 +132,6 @@ class AbstractBatchSampler(Sampler):
         self.np_rng = np.random.RandomState(self.rng)
         torch.manual_seed(self.rng)  # Set torch seed
 
-        # Data augmentation for streamlines:
-        self.noise_gaussian_size = noise_gaussian_size
-        self.noise_gaussian_variability = noise_gaussian_variability
-        self.split_ratio = split_ratio
-        self.reverse_ratio = reverse_ratio
-        if self.split_ratio and not 0 <= self.split_ratio <= 1:
-            raise ValueError('Split ratio must be a float between 0 and 1.')
-        if self.reverse_ratio and not 0 <= self.reverse_ratio <= 1:
-            raise ValueError('Reverse ration must be a float between 0 and 1')
-
         # Batch sampler's logging level can be changed separately from main
         # scripts.
         self.logger = logging.getLogger('batch_sampler_logger')
@@ -236,10 +158,6 @@ class AbstractBatchSampler(Sampler):
             'nb_subjects_per_batch': self.nb_subjects_per_batch,
             'cycles': self.cycles,
             'type': type(self),
-            'noise_gaussian_size': self.noise_gaussian_size,
-            'noise_gaussian_variability': self.noise_gaussian_variability,
-            'reverse_ratio': self.reverse_ratio,
-            'split_ratio': self.split_ratio,
             'step_size': self.step_size,
             'compress': self.compress,
         }
@@ -545,348 +463,3 @@ class AbstractBatchSampler(Sampler):
             # Should be equal to
             # nb_points = lengths_mm / self.dataset.step_size
         return nb_points
-
-    def load_batch(self, streamline_ids_per_subj: List[Tuple[int, list]]) \
-            -> Union[Tuple[List, Dict], Tuple[List, List, List]]:
-        """
-        Fetches the chosen streamlines for all subjects in batch.
-        Pocesses data augmentation.
-
-        Torch uses this function to process the data with the dataloader
-        parallel workers (on cpu). To be used as collate_fn.
-
-        Parameters
-        ----------
-        streamline_ids_per_subj: List[Tuple[int, list]]
-            The list of streamline ids for each subject (relative ids inside
-            each subject's tractogram) for this batch.
-
-        Returns
-        -------
-            (batch_streamlines, final_s_ids_per_subj)
-        Where
-            - batch_streamlines: list[np.array]
-                The new streamlines after data augmentation
-            - final_s_ids_per_subj: Dict[int, slice]
-                The new streamline ids per subj in this augmented batch.
-        """
-        self.logger.debug("        Loading a batch of streamlines!")
-
-        (batch_streamlines, final_s_ids_per_subj) = \
-            self.streamlines_data_preparation(streamline_ids_per_subj)
-
-        return batch_streamlines, final_s_ids_per_subj
-
-    def streamlines_data_preparation(
-            self, streamline_ids_per_subj: List[Tuple[int, list]]):
-        """
-        Parameters
-        -----------
-        streamline_ids_per_subj: List[Tuple[int, list]]
-            The list of streamlines for this batch, per subject (corresponding
-            to the streamlines ids in each subject's tractogram).
-
-        Returns
-        -------
-        batch_streamlines: List[np.array]
-            The new streamlines after data augmentation
-        final_s_ids_per_subj: Dict[int, slice]
-            The new streamline ids per subj in this augmented batch.
-        """
-        # The batch's streamline ids will change throughout processing because
-        # of data augmentation, so we need to do it subject by subject to
-        # keep track of the streamline ids. These final ids will correspond to
-        # the loaded, processed streamlines, not to the ids in the hdf5 file.
-        final_s_ids_per_subj = defaultdict(slice)
-        batch_streamlines = []
-        for subj, s_ids in streamline_ids_per_subj:
-            self.logger.debug("        => Subj: {}".format(subj))
-
-            self.logger.debug(
-                "          Processing data preparation for streamlines ids:\n"
-                "{}".format(s_ids))
-
-            subj_data = self.dataset.subjs_data_list.open_handle_and_getitem(
-                subj)
-            subj_sft_data = subj_data.sft_data_list[self.streamline_group_idx]
-
-            # Get streamlines as sft
-            sft = subj_sft_data.as_sft(s_ids)
-
-            # Resampling streamlines to a fixed step size, if any
-            self.logger.debug("            Resampling: {}"
-                              .format(self.step_size))
-            if self.step_size:
-                if self.dataset.step_size == self.step_size:
-                    self.logger.debug("Step size is the same as when creating "
-                                      "the hdf5 dataset. Not resampling "
-                                      "again.")
-                else:
-                    sft = resample_streamlines_step_size(
-                        sft, step_size=self.step_size)
-
-            # Compressing, if wanted.
-            self.logger.debug(
-                "            Compressing: {}".format(self.compress))
-            if self.compress:
-                sft = compress_sft(sft)
-
-            # Adding noise to coordinates
-            # Noise is considered in mm so we need to make sure the sft is in
-            # rasmm space
-            add_noise = bool(self.noise_gaussian_size and
-                             self.noise_gaussian_size > 0)
-            self.logger.debug("            Adding noise: {}".format(add_noise))
-            if add_noise:
-                sft.to_rasmm()
-                sft = add_noise_to_streamlines(sft,
-                                               self.noise_gaussian_size,
-                                               self.noise_gaussian_variability,
-                                               self.np_rng, self.step_size)
-
-            # Splitting streamlines
-            # This increases the batch size, but does not change the total
-            # length
-            do_split = bool(self.split_ratio and self.split_ratio > 0)
-            self.logger.debug("            Splitting: {}".format(do_split))
-            if do_split:
-                all_ids = np.arange(len(sft))
-                n_to_split = int(np.floor(len(sft) * self.split_ratio))
-                split_ids = self.np_rng.choice(all_ids, size=n_to_split,
-                                               replace=False)
-                sft = split_streamlines(sft, self.np_rng, split_ids)
-
-            # Reversing streamlines
-            do_reverse = self.reverse_ratio and self.reverse_ratio > 0
-            self.logger.debug("            Reversing: {}".format(do_reverse))
-            if do_reverse:
-                ids = np.arange(len(sft))
-                self.np_rng.shuffle(ids)
-                reverse_ids = ids[:int(len(ids) * self.reverse_ratio)]
-                sft = reverse_streamlines(sft, reverse_ids)
-
-            # In case user wants to do more with its data.
-            sft = self.project_specific_data_augmentation(sft)
-
-            # Remember the indices of this subject's (augmented) streamlines
-            ids_start = len(batch_streamlines)
-            ids_end = ids_start + len(sft)
-            final_s_ids_per_subj[subj] = slice(ids_start, ids_end)
-
-            # Add all (augmented) streamlines to the batch
-            # What we want is the streamline coordinates, to eventually get
-            # the underlying input(s). Sending to vox and to corner to
-            # be able to use our trilinear interpolation
-            sft.to_vox()
-            sft.to_corner()
-            batch_streamlines.extend(sft.streamlines)
-
-            return batch_streamlines, final_s_ids_per_subj
-
-    def project_specific_data_augmentation(self, sft: StatefulTractogram):
-        """Please override in your child class if you want to do more than
-        - reversing
-        - adding noise
-        - splitting."""
-        self.logger.debug("            Project-specific data augmentation, if "
-                          "any...")
-
-        return sft
-
-
-class BatchStreamlinesSamplerOneInput(AbstractBatchSampler):
-    """
-    Samples:
-        input = one volume group
-                (data underlying each point of the streamline)
-                (possibly with its neighborhood)
-        target = the whole streamlines as sequences.
-    """
-    def __init__(self, dataset: MultisubjectSubset,
-                 streamline_group_name: str, max_batch_size: int,
-                 batch_size_unit: str, max_chunk_size: int,
-                 rng: int,
-                 nb_subjects_per_batch: int, cycles: int, compress: bool,
-                 step_size: float, split_ratio: float,
-                 noise_gaussian_size: float,
-                 noise_gaussian_variability: float,
-                 reverse_ratio: float, input_group_name, wait_for_gpu: bool,
-                 neighborhood_points: np.ndarray):
-        """
-        Additional parameters compared to super:
-        --------
-        input_group_name: str
-            Name of the input group in the hdf5 dataset.
-        wait_for_gpu: bool
-            If true, will not compute the inputs directly when using
-            load_batch. User can call the compute_inputs method himself later
-            on. Typically, Dataloader (who call load_batch) uses CPU.
-        neighborhood_points: np.ndarray
-            The list of neighborhood points (does not contain 0,0,0 point)
-        """
-        super().__init__(dataset, streamline_group_name, max_batch_size,
-                         batch_size_unit, max_chunk_size, rng,
-                         nb_subjects_per_batch, cycles, step_size, compress,
-                         split_ratio, noise_gaussian_size,
-                         noise_gaussian_variability, reverse_ratio)
-
-        # toDo. Would be more logical to send this as params when using
-        #  load_batch as collate_fn in the Dataloader during training.
-        #  Possible?
-        self.wait_for_gpu = wait_for_gpu
-
-        self.input_group_name = input_group_name
-        self.neighborhood_points = neighborhood_points
-
-        # Find group index in the data_source
-        idx = self.dataset.volume_groups.index(input_group_name)
-        self.input_group_idx = idx
-
-    @property
-    def params(self):
-        p = super().params
-        p.update({
-            'input_group_name': self.input_group_name,
-            'neighborhood_points': self.neighborhood_points,
-            'wait_for_gpu': self.wait_for_gpu
-        })
-        return p
-
-    def load_batch(self, streamline_ids_per_subj: List[Tuple[int, list]],
-                   save_batch_input_mask: bool = False) \
-            -> Union[Tuple[List, Dict],
-                     Tuple[List, List, List]]:
-        """
-        Same as super but also interpolated the underlying inputs (if
-        wanted) for all subjects in batch.
-
-        Torch uses this function to process the data with the dataloader
-        workers. To be used as collate_fn. This part is ran on CPU.
-
-        With self.wait_for_gpu option: avoiding non-necessary operations in the
-        batch sampler, computed on cpu: compute_inputs can be called later by
-        the user.
-        >> inputs = sampler.compute_inputs(batch_streamlines,
-                                           streamline_ids_per_subj)
-
-        Additional parameters compared to super:
-        ----------
-        save_batch_input_mask: bool
-            Debugging purposes. Saves the input coordinates as a mask. Must be
-            used together with wait_for_gpu=False. The inputs will be modified
-            to a tuple containing the batch_streamlines (with wait_for_gpu set
-            to False, the directions only would be returned), to compare the
-            streamlines with masks.
-
-        Returns
-        -------
-        If self.wait_for_gpu: same as super. Else, also returns
-            batch_inputs : List of torch.Tensor
-                Inputs volume loaded from the given group name. Data is
-                flattened. Length of the list is the number of streamlines.
-                Size of tensor at index i is [N_i-1, nb_features].
-        """
-        batch = super().load_batch(streamline_ids_per_subj)
-
-        if self.wait_for_gpu:
-            # Only returning the streamlines for now.
-            # Note that this is the same as using data_preparation_cpu_step
-            self.logger.debug("            Not loading the input data because "
-                              "user prefers to do it later on GPU.")
-
-            return batch
-        else:
-            # At this point batch_streamlines are in voxel space with
-            # corner origin.
-            batch_streamlines, final_s_ids_per_subj = batch
-
-            # Get the inputs
-            self.logger.debug("        Loading a batch of inputs!")
-            batch_inputs = self.compute_inputs(batch_streamlines,
-                                               final_s_ids_per_subj,
-                                               save_batch_input_mask)
-
-            return batch_streamlines, final_s_ids_per_subj, batch_inputs
-
-    def compute_inputs(self, batch_streamlines: List[np.ndarray],
-                       streamline_ids_per_subj: Dict[int, slice],
-                       save_batch_input_mask: bool = False,
-                       device=torch.device('cpu')):
-        """
-        Get the DWI (depending on volume: as raw, SH, fODF, etc.) volume for
-        each point in each streamline (+ depending on options: neighborhood,
-        and preceding diretion)
-
-        Params
-        ------
-        batch_streamlines: list[np.array]
-            The streamlines (after data augmentation) in voxel space, with
-            corner origin.
-        streamline_ids_per_subj: Dict[int, slice]
-            The ids corresponding to each subject (so we can load the
-            associated subject's input volume).
-        save_batch_input_mask: bool
-            Debugging purposes. Saves the input coordinates as a mask. Must be
-            used together with wait_for_gpu=False. The inputs will be modified
-            to a tuple containing the batch_streamlines (with wait_for_gpu set
-            to False, the directions only would be returned), to compare the
-            streamlines with masks.
-        device: torch device
-            Torch device.
-
-        Returns
-        -------
-        batch_x_data : List
-            The list of (list of) inputs for each streamlines
-        """
-        batch_x_data = []
-        batch_input_masks = []
-
-        for subj, y_ids in streamline_ids_per_subj.items():
-            # Flatten = concatenate signal for all streamlines to process
-            # faster. We don't use the last coord because it is used only to
-            # compute the last target direction, it's not really an input
-            flat_subj_x_coords = np.concatenate(
-                [s[:-1] for s in batch_streamlines[y_ids]], axis=0)
-
-            # Getting the subject's volume and sending to CPU/GPU
-            # If data is lazy, get volume from cache or send to cache if
-            # it wasn't there yet.
-            data_tensor = self.dataset.get_volume_verify_cache(
-                subj, self.input_group_idx, device=device, non_blocking=True)
-
-            # Prepare the volume data, possibly adding neighborhood
-            # (Thus new coords_torch possibly contain the neighborhood points)
-            # Coord_clipped contain the coords after interpolation
-            # Trilinear interpolation uses origin=corner, vox space, but ok
-            # because in load_batch, we use sft.to_vox and sft.to_corner
-            # before adding streamline to batch.
-            subj_x_data, coords_torch = interpolate_volume_in_neighborhood(
-                data_tensor, flat_subj_x_coords, self.neighborhood_points,
-                device)
-
-            # Split the flattened signal back to streamlines
-            lengths = [len(s) - 1 for s in batch_streamlines[y_ids]]
-            subbatch_x_data = subj_x_data.split(lengths)
-            batch_x_data.extend(subbatch_x_data)
-
-            if save_batch_input_mask:
-                print("DEBUGGING MODE. Returning batch_streamlines "
-                      "and mask together with inputs.")
-
-                # Clipping used coords (i.e. possibly with neighborhood)
-                # outside volume
-                lower = torch.as_tensor([0, 0, 0], device=device)
-                upper = torch.as_tensor(data_tensor.shape[:3], device=device)
-                upper -= 1
-                coords_to_idx_clipped = torch.min(
-                    torch.max(torch.floor(coords_torch).long(), lower),
-                    upper)
-                input_mask = torch.tensor(np.zeros(data_tensor.shape[0:3]))
-                for s in range(len(coords_torch)):
-                    input_mask.data[tuple(coords_to_idx_clipped[s, :])] = 1
-                batch_input_masks.append(input_mask)
-
-                return batch_input_masks, batch_x_data
-
-        return batch_x_data
