+++ please_copy_and_adapt/resume_training_from_checkpoint.py
@@ -6,7 +6,7 @@ import os
 from os import path
 
 from dwi_ml.data.dataset.utils import prepare_multisubjectdataset
-from dwi_ml.training.utils.batch_sampler import prepare_batchsamplers_oneinput
+from dwi_ml.training.utils.batch_samplers import prepare_batchsamplers_train_valid
 from dwi_ml.experiment_utils.timer import Timer
 from dwi_ml.models.main_models import MainModelAbstract
 from dwi_ml.training.utils.trainer import run_experiment
@@ -72,8 +72,8 @@ def init_from_checkpoint(args):
     args_va_s = None if checkpoint_state['valid_sampler_params'] is None else \
         argparse.Namespace(**checkpoint_state['valid_sampler_params'])
     training_batch_sampler, validation_batch_sampler = \
-        prepare_batchsamplers_oneinput(dataset, args_tr_s, args_va_s,
-                                       model.neighborhood_points)
+        prepare_batchsamplers_train_valid(dataset, args_tr_s, args_va_s,
+                                          model.neighborhood_points)
 
     # Instantiate trainer
     with Timer("\n\nPreparing trainer", newline=True, color='red'):
